Make traversal order pre-order in bufferization.
Make the traversal order pre-order since for ops with nested regions,
it is necessary to process the op before the region.
Changing to pre-order also means that regions of cloned ops dont get
processed during the work to convert tensor to memrefs. As a result,
move the conversion of ops that are just using the tensor value, and
not creating a tensor to be processed in a separate walk.